raflow: WIP migrating struct definitions and methods to rlib
