+++ rlib/raflow.go
@@ -1,5 +1,14 @@
 package rlib
 
+import (
+	"bytes"
+	"context"
+	"encoding/json"
+	"fmt"
+	"reflect"
+	"time"
+)
+
 // RAFlow etc.. all are list of all flows exist in the system
 const (
 	RAFlow string = "RA"
@@ -57,4 +66,967 @@ var RAFlowPartsMap = Str2Int64Map{
 	"rentables":   int64(RentablesRAFlowPart),
 	"parentchild": int64(ParentChildRAFlowPart),
 	"tie":         int64(TieRAFlowPart),
+}
+
+// RAFlowJSONData holds the struct for all the parts being involed in rental agreement flow
+type RAFlowJSONData struct {
+	Dates       RADatesFlowData         `json:"dates"`
+	People      []RAPeopleFlowData      `json:"people"`
+	Pets        []RAPetsFlowData        `json:"pets"`
+	Vehicles    []RAVehiclesFlowData    `json:"vehicles"`
+	Rentables   []RARentablesFlowData   `json:"rentables"`
+	ParentChild []RAParentChildFlowData `json:"parentchild"`
+	Tie         RATieFlowData           `json:"tie"`
+	Meta        RAFlowMetaInfo          `json:"meta"`
+}
+
+// RAFlowMetaInfo holds meta info about a rental agreement flow data
+type RAFlowMetaInfo struct {
+	RAID         int64 // 0 = it's new, >0 = existing one
+	LastTMPPETID int64
+	LastTMPVID   int64
+	LastTMPTCID  int64
+	LastTMPASMID int64
+	HavePets     bool
+	HaveVehicles bool
+	RAFLAGS      int64
+}
+
+// RADatesFlowData contains data in the dates part of RA flow
+type RADatesFlowData struct {
+	BID             int64    `validate:"number,min=1"`
+	AgreementStart  JSONDate `validate:"date"` // TermStart
+	AgreementStop   JSONDate `validate:"date"` // TermStop
+	RentStart       JSONDate `validate:"date"`
+	RentStop        JSONDate `validate:"date"`
+	PossessionStart JSONDate `validate:"date"`
+	PossessionStop  JSONDate `validate:"date"`
+}
+
+// RAPeopleFlowData contains data in the background-info part of RA flow
+type RAPeopleFlowData struct {
+	TMPTCID int64 `validate:"number,min=1"`
+	BID     int64 `validate:"number,min=1"`
+	TCID    int64 `validate:"number,min=1"`
+
+	// Role
+	IsRenter    bool `validate:"-"`
+	IsOccupant  bool `validate:"-"`
+	IsGuarantor bool `validate:"-"`
+
+	// ---------- Basic Info -----------
+	FirstName      string `validate:"string,min=1,max=100"`
+	MiddleName     string `validate:"string,min=1,max=100"`
+	LastName       string `validate:"string,min=1,max=100"`
+	PreferredName  string `validate:"string,min=1,max=100,omitempty"`
+	IsCompany      bool   `validate:"number,min=1,max=1"`
+	CompanyName    string `validate:"string,min=1,max=100"`
+	PrimaryEmail   string `validate:"email,omitempty"`
+	SecondaryEmail string `validate:"email,omitempty"`
+	WorkPhone      string `validate:"number,min=1,max=100,omitempty"`
+	CellPhone      string `validate:"number,min=1,max=100,omitempty"`
+	Address        string `validate:"string,min=1,max=100"`
+	Address2       string `validate:"string,min=0,max=100,omitempty"`
+	City           string `validate:"string,min=1,max=100"`
+	State          string `validate:"string,min=1,max=25"`
+	PostalCode     string `validate:"number,min=1,max=100"`
+	Country        string `validate:"string,min=1,max=100"`
+	Website        string `validate:"string,min=1,max=100,omitempty"`
+	Comment        string `validate:"string,min=1,max=2048,omitempty"`
+
+	// ---------- Prospect -----------
+	CompanyAddress    string `validate:"string,min=1,max=100"`
+	CompanyCity       string `validate:"string,min=1,max=100"`
+	CompanyState      string `validate:"string,min=1,max=100"`
+	CompanyPostalCode string `validate:"number,min=1,max=100"`
+	CompanyEmail      string `validate:"email"`
+	CompanyPhone      string `validate:"number,min=1,max=100"`
+	Occupation        string `validate:"string,min=1,max=100"`
+
+	// Current Address information
+	CurrentAddress           string `validate:"string,min=1,max=100"`
+	CurrentLandLordName      string `validate:"string,min=1,max=100"`
+	CurrentLandLordPhoneNo   string `validate:"number,min=1"`
+	CurrentLengthOfResidency string `validate:"string,min=1,max=100"`
+	CurrentReasonForMoving   int64  `validate:"number,min=1"` // Reason for moving
+
+	// Prior Address information
+	PriorAddress           string `validate:"string,min=1,max=100"`
+	PriorLandLordName      string `validate:"string,min=1,max=100"`
+	PriorLandLordPhoneNo   string `validate:"number,min=1"`
+	PriorLengthOfResidency string `validate:"string,min=1,max=100"`
+	PriorReasonForMoving   int64  `validate:"number,min=1"` // Reason for moving
+
+	// Have you ever been
+	Evicted          bool   `validate:"-"` // Evicted
+	EvictedDes       string `validate:"string,min=1,max=2048,omitempty"`
+	Convicted        bool   `validate:"-"` // Arrested or convicted of a Convicted
+	ConvictedDes     string `validate:"string,min=1,max=2048,omitempty"`
+	Bankruptcy       bool   `validate:"-"` // Declared Bankruptcy
+	BankruptcyDes    string `validate:"string,min=1,max=2048,omitempty"`
+	OtherPreferences string `validate:"string,min=1,max=1024"`
+	//FollowUpDate             JSONDate
+	//CommissionableThirdParty string
+	SpecialNeeds string `validate:"string,min=1,max=1024"` // In an effort to accommodate you, please advise us of any special needs
+
+	// ---------- Payor -----------
+	CreditLimit         float64 `validate:"number:float,min=0.10"`
+	TaxpayorID          string  `validate:"string,min=1,max=25"`
+	GrossIncome         float64 `validate:"number:float,min=0.10"`
+	SSN                 string  `validate:"string,min=1,max=128"`
+	DriversLicense      string  `validate:"string,min=1,max=128"`
+	ThirdPartySource    int64   `validate:"number,min=1,omitempty"`
+	EligibleFuturePayor bool    `validate:"-"`
+
+	// ---------- User -----------
+	Points      int64    `validate:"number,min=1,omitempty"`
+	DateofBirth JSONDate `validate:"date"`
+	// Emergency contact information
+	EmergencyContactName      string `validate:"string,min=1,max=100"`
+	EmergencyContactAddress   string `validate:"string,min=1,max=100"`
+	EmergencyContactTelephone string `validate:"number,min=1,max=100"`
+	EmergencyContactEmail     string `validate:"email"`
+	AlternateAddress          string `validate:"string,min=1,max=100"`
+	EligibleFutureUser        bool   `validate:"number,min=1"`
+	Industry                  string `validate:"string,min=1,max=100"`
+	SourceSLSID               int64  `validate:"number,min=1"`
+}
+
+// RAPetsFlowData contains data in the pets part of RA flow
+type RAPetsFlowData struct {
+	TMPPETID int64        `validate:"number,min=1"`
+	BID      int64        `validate:"number,min=1"`
+	PETID    int64        `validate:"number,min=1"`
+	TMPTCID  int64        `validate:"number,min=1"`
+	Name     string       `validate:"string,min=1,max=100"`
+	Type     string       `validate:"string,min=1,max=100"`
+	Breed    string       `validate:"string,min=1,max=100"`
+	Color    string       `validate:"string,min=1,max=100"`
+	Weight   float64      `validate:"number:float,min=0.0"`
+	DtStart  JSONDate     `validate:"date"`
+	DtStop   JSONDate     `validate:"date"`
+	Fees     []RAFeesData `validate:"-"`
+}
+
+// RAVehiclesFlowData contains data in the vehicles part of RA flow
+type RAVehiclesFlowData struct {
+	TMPVID              int64        `validate:"number,min=1"`
+	BID                 int64        `validate:"number,min=1"`
+	VID                 int64        `validate:"number,min=1"`
+	TMPTCID             int64        `validate:"number,min=1"`
+	VIN                 string       `validate:"string,min=1"`
+	VehicleType         string       `validate:"string,min=1,max=80"`
+	VehicleMake         string       `validate:"string,min=1,max=80"`
+	VehicleModel        string       `validate:"string,min=1,max=80"`
+	VehicleColor        string       `validate:"string,min=1,max=80"`
+	VehicleYear         int64        `validate:"number,min=1"`
+	LicensePlateState   string       `validate:"string,min=1,max=80"`
+	LicensePlateNumber  string       `validate:"string,min=1,max=80"`
+	ParkingPermitNumber string       `validate:"string,min=1,max=80"`
+	DtStart             JSONDate     `validate:"date"`
+	DtStop              JSONDate     `validate:"date"`
+	Fees                []RAFeesData `validate:"-"`
+}
+
+// RARentablesFlowData contains data in the rentables part of RA flow
+type RARentablesFlowData struct {
+	BID             int64 `validate:"number,min=1"`
+	RID             int64 `validate:"number,min=1"`
+	RTID            int64 `validate:"number,min=1"`
+	RTFLAGS         uint64
+	RentableName    string  `validate:"string,min=1,max=100"`
+	RentCycle       int64   `validate:"number,min=1"`
+	AtSigningPreTax float64 `validate:"number:float,min=0.00"`
+	SalesTax        float64 `validate:"number:float,min=0.00"`
+	// SalesTaxAmt    float64 // FUTURE RELEASE
+	TransOccTax float64 `validate:"number:float,min=0.00"`
+	// TransOccAmt    float64 // FUTURE RELEASE
+	Fees []RAFeesData `validate:"-"`
+}
+
+// RAFeesData struct used for pet, vehicles, rentable fees
+type RAFeesData struct {
+	TMPASMID        int64    `validate:"number,min=1"` // unique ID to manage fees uniquely across all fees in raflow json data
+	ASMID           int64    `validate:"number,min=0"` // the permanent table assessment id if it is an existing RAID
+	ARID            int64    `validate:"number,min=1"`
+	ARName          string   `validate:"string,min=1,max=100"`
+	ContractAmount  float64  `validate:"number:float,min=0.00"`
+	RentCycle       int64    `validate:"number,min=1"`
+	Start           JSONDate `validate:"date"`
+	Stop            JSONDate `validate:"date"`
+	AtSigningPreTax float64  `validate:"number:float,min=0.00"`
+	SalesTax        float64  `validate:"number:float,min=0.00"`
+	// SalesTaxAmt     float64       // FUTURE RELEASE
+	TransOccTax float64 `validate:"number:float,min=0.00"`
+	// TransOccAmt float64 // FUTURE RELEASE
+	Comment string `validate:"-"`
+}
+
+// RAParentChildFlowData contains data in the Parent/Child part of RA flow
+type RAParentChildFlowData struct {
+	BID  int64 `validate:"number,min=1"`
+	PRID int64 `validate:"number,min=1"` // parent rentable ID
+	CRID int64 `validate:"number,min=1"` // child rentable ID
+}
+
+// RATieFlowData contains data in the tie part of RA flow
+type RATieFlowData struct {
+	People []RATiePeopleData `json:"people"`
+}
+
+// RATiePeopleData holds data from tie section for a payor to a rentable
+type RATiePeopleData struct {
+	BID     int64 `validate:"number,min=1"`
+	PRID    int64 `validate:"number,min=1"`
+	TMPTCID int64 `validate:"number,min=1"` // user's temp json record reference id
+}
+
+// GetRAFlowPartDataFromJSON returns json data in bytes
+// in structure of a raflow part from client data
+func GetRAFlowPartDataFromJSON(BID int64, data json.RawMessage, partType int, flow *Flow) ([]byte, []byte, error) {
+
+	var (
+		modFlowPartData = []byte(nil)
+		modMetaData     = []byte(nil)
+		err             error
+		raFlowData      RAFlowJSONData
+	)
+
+	// get the whole raflow data from Flow type data
+	err = json.Unmarshal(flow.Data, &raFlowData)
+	if err != nil {
+		// if it's an error then return with nil data
+		return modMetaData, modFlowPartData, err
+	}
+
+	// JSON Marshal with address
+	// REF: https://stackoverflow.com/questions/21390979/custom-marshaljson-never-gets-called-in-go
+
+	// is it blank string or null json data
+	isBlankJSONData := bytes.Equal([]byte(data), []byte(``)) || bytes.Equal([]byte(data), []byte(`null`))
+
+	switch RAFlowPartType(partType) {
+	case DatesRAFlowPart:
+		a := RADatesFlowData{}
+
+		// if the struct provided with some data then check it for
+		// json validation
+		if !(isBlankJSONData) {
+			err := json.Unmarshal(data, &a)
+			if err != nil {
+				// if it's an error then return with nil data
+				return modMetaData, modFlowPartData, err
+			}
+		} else {
+			// it's null/blank data then initialize with default data
+			currentDateTime := time.Now()
+			nextYearDateTime := currentDateTime.AddDate(1, 0, 0)
+
+			a.BID = BID
+			a.RentStart = JSONDate(currentDateTime)
+			a.RentStop = JSONDate(nextYearDateTime)
+			a.AgreementStart = JSONDate(currentDateTime)
+			a.AgreementStop = JSONDate(nextYearDateTime)
+			a.PossessionStart = JSONDate(currentDateTime)
+			a.PossessionStop = JSONDate(nextYearDateTime)
+		}
+
+		// json marshalled for struct
+		modFlowPartData, err = json.Marshal(&a)
+
+	case PeopleRAFlowPart:
+		a := []RAPeopleFlowData{}
+
+		// if the struct provided with some data then check it for
+		// json validation
+		if !(isBlankJSONData) {
+			err := json.Unmarshal(data, &a)
+
+			// auto assign TMPTCID
+			for i := range a {
+				if a[i].TMPTCID == 0 { // if zero then assign new from last saved ID
+					raFlowData.Meta.LastTMPTCID++
+					a[i].TMPTCID = raFlowData.Meta.LastTMPTCID
+				}
+			}
+
+			if err != nil {
+				// if it's an error then return with nil data
+				return modMetaData, modFlowPartData, err
+			}
+		}
+
+		// json marshalled for struct
+		modFlowPartData, err = json.Marshal(&a)
+
+	case PetsRAFlowPart:
+		a := []RAPetsFlowData{}
+
+		// if the struct provided with some data then check it for
+		// json validation
+		if !(isBlankJSONData) {
+			err := json.Unmarshal(data, &a)
+
+			// auto assign TMPPETID
+			for i := range a {
+				// If Fees not initialized then
+				if len(a[i].Fees) == 0 {
+					a[i].Fees = []RAFeesData{}
+				}
+
+				if a[i].TMPPETID == 0 { // if zero then assign new from last saved ID
+					raFlowData.Meta.LastTMPPETID++
+					a[i].TMPPETID = raFlowData.Meta.LastTMPPETID
+
+					// manage TMPASMID in Fees
+					for j := range a[i].Fees {
+						if a[i].Fees[j].TMPASMID == 0 {
+							raFlowData.Meta.LastTMPASMID++
+							a[i].Fees[j].TMPASMID = raFlowData.Meta.LastTMPASMID
+						}
+					}
+				}
+			}
+
+			// Update HavePets flag in meta information
+			raFlowData.Meta.HavePets = len(a) > 0
+
+			if err != nil {
+				// if it's an error then return with nil data
+				return modMetaData, modFlowPartData, err
+			}
+		}
+
+		// json marshalled for struct
+		modFlowPartData, err = json.Marshal(&a)
+
+	case VehiclesRAFlowPart:
+		a := []RAVehiclesFlowData{}
+
+		// if the struct provided with some data then check it for
+		// json validation
+		if !(isBlankJSONData) {
+			err := json.Unmarshal(data, &a)
+
+			// auto assign TMPVID
+			for i := range a {
+				// If Fees not initialized then
+				if len(a[i].Fees) == 0 {
+					a[i].Fees = []RAFeesData{}
+				}
+
+				if a[i].TMPVID == 0 { // if zero then assign new from last saved ID
+					raFlowData.Meta.LastTMPVID++
+					a[i].TMPVID = raFlowData.Meta.LastTMPVID
+
+					// manage TMPASMID in fees
+					for j := range a[i].Fees {
+						if a[i].Fees[j].TMPASMID == 0 {
+							raFlowData.Meta.LastTMPASMID++
+							a[i].Fees[j].TMPASMID = raFlowData.Meta.LastTMPASMID
+						}
+					}
+				}
+			}
+
+			// Update HaveVehicles flag in meta information
+			raFlowData.Meta.HaveVehicles = len(a) > 0
+
+			if err != nil {
+				// if it's an error then return with nil data
+				return modMetaData, modFlowPartData, err
+			}
+		}
+
+		// json marshalled for struct
+		modFlowPartData, err = json.Marshal(&a)
+
+	case RentablesRAFlowPart:
+		a := []RARentablesFlowData{}
+
+		// if the struct provided with some data then check it for
+		// json validation
+		if !(isBlankJSONData) {
+			err := json.Unmarshal(data, &a)
+
+			for i := range a {
+
+				// If Fees not initialized then
+				if len(a[i].Fees) == 0 {
+					a[i].Fees = []RAFeesData{}
+				}
+
+				// manage TMPASMID in fees
+				for j := range a[i].Fees {
+					if a[i].Fees[j].TMPASMID == 0 {
+						raFlowData.Meta.LastTMPASMID++
+						a[i].Fees[j].TMPASMID = raFlowData.Meta.LastTMPASMID
+					}
+				}
+
+			}
+
+			if err != nil {
+				// if it's an error then return with nil data
+				return modMetaData, modFlowPartData, err
+			}
+		}
+
+		// json marshalled for struct
+		modFlowPartData, err = json.Marshal(&a)
+
+	case ParentChildRAFlowPart:
+		a := []RAParentChildFlowData{}
+
+		// if the struct provided with some data then check it for
+		// json validation
+		if !(isBlankJSONData) {
+			err := json.Unmarshal(data, &a)
+			if err != nil {
+				// if it's an error then return with nil data
+				return modMetaData, modFlowPartData, err
+			}
+		}
+
+		// json marshalled for struct
+		modFlowPartData, err = json.Marshal(&a)
+
+	case TieRAFlowPart:
+		a := RATieFlowData{}
+
+		// if the struct provided with some data then check it for
+		// json validation
+		if !(isBlankJSONData) {
+			err := json.Unmarshal(data, &a)
+
+			// check for each sliced data field
+			// if it's blank then initialize it
+			if len(a.People) == 0 {
+				a.People = []RATiePeopleData{}
+			}
+
+			if err != nil {
+				// if it's an error then return with nil data
+				return modMetaData, modFlowPartData, err
+			}
+		}
+
+		// json marshalled for struct
+		modFlowPartData, err = json.Marshal(&a)
+
+	default:
+		err = fmt.Errorf("unrecognized part type in RA flow: %d", partType)
+	}
+
+	// if error occured in above switch cases execution
+	// while marshaling content in json then only
+	if err != nil {
+		return modMetaData, modFlowPartData, err
+	}
+
+	// now marshal json data back to raflow
+	modMetaData, err = json.Marshal(&raFlowData.Meta)
+	if err != nil {
+		// if it's an error then return with nil data
+		return modMetaData, modFlowPartData, err
+	}
+
+	// finally return with modified data
+	return modMetaData, modFlowPartData, err
+}
+
+// InsertInitialRAFlow writes a bunch of flow's sections record for a particular RA
+func InsertInitialRAFlow(ctx context.Context, BID, UID int64) (int64, error) {
+
+	var (
+		flowID int64
+		err    error
+	)
+
+	// current date and next year date
+	currentDateTime := time.Now()
+	nextYearDateTime := currentDateTime.AddDate(1, 0, 0)
+
+	// rental agreement flow data
+	initialRAFlow := RAFlowJSONData{
+		Dates: RADatesFlowData{
+			BID:             BID,
+			RentStart:       JSONDate(currentDateTime),
+			RentStop:        JSONDate(nextYearDateTime),
+			AgreementStart:  JSONDate(currentDateTime),
+			AgreementStop:   JSONDate(nextYearDateTime),
+			PossessionStart: JSONDate(currentDateTime),
+			PossessionStop:  JSONDate(nextYearDateTime),
+		},
+		People:      []RAPeopleFlowData{},
+		Pets:        []RAPetsFlowData{},
+		Vehicles:    []RAVehiclesFlowData{},
+		Rentables:   []RARentablesFlowData{},
+		ParentChild: []RAParentChildFlowData{},
+		Tie: RATieFlowData{
+			People: []RATiePeopleData{},
+		},
+	}
+
+	// get json marshelled byte data for above struct
+	raflowJSONData, err := json.Marshal(&initialRAFlow)
+	if err != nil {
+		Ulog("Error while marshalling json data of initialRAFlow: %s\n", err.Error())
+		return flowID, err
+	}
+
+	// initial Flow struct
+	Console("New Flow\n")
+	a := Flow{
+		BID:       BID,
+		FlowID:    0, // it's new flowID,
+		UserRefNo: GenerateUserRefNo(),
+		FlowType:  RAFlow,
+		Data:      raflowJSONData,
+		CreateBy:  UID,
+		LastModBy: UID,
+	}
+
+	Console("New flow UserRefNo = %s\n", a.UserRefNo)
+
+	// insert new flow
+	flowID, err = InsertFlow(ctx, &a)
+	if err != nil {
+		Ulog("Error while inserting Flow: %s\n", err.Error())
+		return flowID, err
+	}
+
+	return flowID, err
+}
+
+// RAFlowDataDiff returns true/false if there is any data
+// difference between raflow json data and permanent table
+// data
+func RAFlowDataDiff(ctx context.Context, RAID int64) (isDiff bool, err error) {
+	const funcname = "RAFlowDataDiff"
+	var (
+		ra                      RentalAgreement
+		flow                    Flow
+		flowData, permanentData RAFlowJSONData
+	)
+
+	// --------------------------------------------------------
+	// GET DATA FROM TEMP FLOW table
+	// --------------------------------------------------------
+	flow, err = GetFlowForRAID(ctx, "RA", RAID)
+	if err != nil {
+		return
+	}
+
+	err = json.Unmarshal(flow.Data, &flowData)
+	if err != nil {
+		return
+	}
+
+	// --------------------------------------------------------
+	// GET PERMANENT RA and GET FLOWDATA FOR IT
+	// --------------------------------------------------------
+	ra, err = GetRentalAgreement(ctx, RAID)
+	if err != nil {
+		return
+	}
+
+	// convert permanent ra to flow data and get it
+	permanentData, err = ConvertRA2Flow(ctx, &ra)
+	if err != nil {
+		return
+	}
+
+	// --------------------------------------------------------
+	// WE DON'T NEED meta DATA TO DIFF BETWEEN BOTH JSON DATA
+	// --------------------------------------------------------
+	permanentData.Meta, flowData.Meta = RAFlowMetaInfo{}, RAFlowMetaInfo{}
+
+	// NOW TAKE THE DIFF USING REFLECT
+	sameData := reflect.DeepEqual(permanentData, flowData) // returns true, if both are equal
+	isDiff = !sameData
+
+	return
+}
+
+// ConvertRA2Flow does all the heavy lifting to convert existing
+// rental agreement data to raflow data
+//
+// INPUTS:
+//     ctx    database context for transactions
+//     ra     the rental agreement to move into a flow
+//
+// RETURNS:
+//     the RAFlowJSONData
+//     any error encountered
+//-------------------------------------------------------------------------
+func ConvertRA2Flow(ctx context.Context, ra *RentalAgreement) (RAFlowJSONData, error) {
+	const funcname = "ConvertRA2Flow"
+	fmt.Printf("Entered in %s\n", funcname)
+
+	//-------------------------------------------------------------
+	// This is the datastructure we need to fill out and save...
+	//-------------------------------------------------------------
+	var raf = RAFlowJSONData{
+		Dates: RADatesFlowData{
+			BID:             ra.BID,
+			RentStart:       JSONDate(ra.RentStart),
+			RentStop:        JSONDate(ra.RentStop),
+			AgreementStart:  JSONDate(ra.AgreementStart),
+			AgreementStop:   JSONDate(ra.AgreementStop),
+			PossessionStart: JSONDate(ra.PossessionStart),
+			PossessionStop:  JSONDate(ra.PossessionStop),
+		},
+		People:      []RAPeopleFlowData{},
+		Pets:        []RAPetsFlowData{},
+		Vehicles:    []RAVehiclesFlowData{},
+		Rentables:   []RARentablesFlowData{},
+		ParentChild: []RAParentChildFlowData{},
+		Tie: RATieFlowData{
+			People: []RATiePeopleData{},
+		},
+		Meta: RAFlowMetaInfo{RAID: ra.RAID},
+	}
+
+	//-------------------------------------------------------------------------
+	// Add Users...
+	//
+	// Note: we need to add users before payors in order to
+	// ensure that all the pets and vehicles are loaded.  This is because
+	// of two behaviors of the code.  First, pets and vehicles are loaded only
+	// when the person is loaded with the User (occupant) flag set.  Second,
+	// a person is not added twice, and if you load a Payor first -- the pets
+	// and vehicles will NOT be loaded then when you call it the second time
+	// to load the same transactant as a User the code will see that the
+	// transactant has already been loaded and return without doing anything
+	// other than setting the Payor (renter) flag.
+	//-------------------------------------------------------------------------
+	n, err := GetAllRentalAgreementRentables(ctx, ra.RAID)
+	if err != nil {
+		return raf, nil
+	}
+	for j := 0; j < len(n); j++ {
+		rulist, err := GetRentableUsersInRange(ctx, n[j].RID, &ra.AgreementStart, &ra.AgreementStop)
+		if err != nil {
+			return raf, nil
+		}
+		for k := 0; k < len(rulist); k++ {
+			addRAPtoFlow(ctx, rulist[k].TCID, n[j].RID, &raf, true, false, true)
+		}
+	}
+
+	//-------------------------------------------------------------------------
+	// Add Payors...
+	//-------------------------------------------------------------------------
+	m, err := GetRentalAgreementPayorsInRange(ctx, ra.RAID, &ra.AgreementStart, &ra.AgreementStop)
+	if err != nil {
+		return raf, nil
+	}
+	for i := 0; i < len(m); i++ {
+		if err = addRAPtoFlow(ctx, m[i].TCID, 0 /*no RID here*/, &raf, true /*check dups*/, true /*renter*/, false); err != nil {
+			return raf, nil
+		}
+	}
+
+	//-------------------------------------------------------------------------
+	// Add Rentables
+	//-------------------------------------------------------------------------
+	now := time.Now()
+	o, err := GetRentalAgreementRentables(ctx, ra.RAID, &ra.AgreementStart, &ra.AgreementStop)
+	if err != nil {
+		return raf, nil
+	}
+	for i := 0; i < len(o); i++ {
+		rnt, err := GetRentable(ctx, o[i].RID)
+		if err != nil {
+			return raf, nil
+		}
+		rtr, err := GetRentableTypeRefForDate(ctx, o[i].RID, &now)
+		if err != nil {
+			return raf, nil
+		}
+		var rt RentableType
+		if err = GetRentableType(ctx, rtr.RTID, &rt); err != nil {
+			return raf, nil
+		}
+		var rfd = RARentablesFlowData{
+			BID:          o[i].BID,
+			RID:          o[i].RID,
+			RTID:         rtr.RTID,
+			RTFLAGS:      rt.FLAGS,
+			RentableName: rnt.RentableName,
+			RentCycle:    rt.RentCycle,
+		}
+
+		//---------------------------------------------------------
+		// Add the assessments associated with the Rentable...
+		// For this we want to load all 1-time fees and all
+		// recurring fees.
+		//---------------------------------------------------------
+		asms, err := GetAssessmentsByRAIDRID(ctx, rfd.BID, ra.RAID, rfd.RID)
+		if err != nil {
+			return raf, nil
+		}
+		for j := 0; j < len(asms); j++ {
+
+			//----------------------------------------------------------
+			// Get the account rule for this assessment...
+			//----------------------------------------------------------
+			ar, err := GetAR(ctx, asms[j].ARID)
+			if err != nil {
+				return raf, nil
+			}
+
+			//----------------------------------------------------------
+			// Handle Rentable Fees that are NOT Pet or Vehicle related
+			//----------------------------------------------------------
+			if ar.FLAGS&(128|256) == 0 {
+				var fee = RAFeesData{
+					ASMID:          asms[j].ASMID,
+					ARID:           asms[j].ARID,
+					ARName:         ar.Name,
+					ContractAmount: asms[j].Amount,
+					RentCycle:      asms[j].RentCycle,
+					Start:          JSONDate(asms[j].Start),
+					Stop:           JSONDate(asms[j].Stop),
+					Comment:        asms[j].Comment,
+				}
+				rfd.Fees = append(rfd.Fees, fee)
+			}
+
+			//----------------------------------------------------------
+			// Handle PET Fees
+			//----------------------------------------------------------
+			// TMPASMID        int64 // unique ID to manage fees uniquely across all fees in raflow json data
+			// ASMID           int64 // the permanent table assessment id if it is an existing RAID
+			// ARID            int64
+			// ARName          string
+			// ContractAmount  float64
+			// RentCycle       int64
+			// Start           JSONDate
+			// Stop            JSONDate
+			// AtSigningPreTax float64
+			// SalesTax        float64
+			if ar.FLAGS&(128) != 0 { // Is it a pet fee?
+				petid := asms[j].AssocElemID // find the pet...
+				for k := 0; k < len(raf.Pets); k++ {
+					if raf.Pets[k].PETID == petid {
+						raf.Meta.LastTMPASMID++
+						var pf = RAFeesData{
+							TMPASMID:       raf.Meta.LastTMPASMID,
+							ARID:           asms[j].ARID,
+							ASMID:          asms[j].ASMID,
+							ARName:         ar.Name,
+							RentCycle:      asms[j].RentCycle,
+							Start:          JSONDate(asms[j].Start),
+							Stop:           JSONDate(asms[j].Stop),
+							ContractAmount: asms[j].Amount,
+							Comment:        asms[j].Comment,
+						}
+						raf.Pets[k].Fees = append(raf.Pets[k].Fees, pf)
+						break
+					}
+				}
+			}
+			//----------------------------------------------------------
+			// Handle VEHICLE Fees
+			//----------------------------------------------------------
+			if ar.FLAGS&(256) != 0 { // Is it a vehicle fee?
+				vid := asms[j].AssocElemID // find the vehicle...
+				for k := 0; k < len(raf.Vehicles); k++ {
+					if raf.Vehicles[k].VID == vid {
+						raf.Meta.LastTMPASMID++
+						var pf = RAFeesData{
+							TMPASMID:       raf.Meta.LastTMPASMID,
+							ARID:           asms[j].ARID,
+							ASMID:          asms[j].ASMID,
+							ARName:         ar.Name,
+							ContractAmount: asms[j].Amount,
+							RentCycle:      asms[j].RentCycle,
+							Start:          JSONDate(asms[j].Start),
+							Stop:           JSONDate(asms[j].Stop),
+							Comment:        asms[j].Comment,
+						}
+						raf.Vehicles[k].Fees = append(raf.Vehicles[k].Fees, pf)
+						break
+					}
+				}
+			}
+		}
+
+		raf.Rentables = append(raf.Rentables, rfd)
+	}
+
+	return raf, nil
+}
+
+// addRAPtoFlow adds a new person to raf.People.  The renter/occupant flags
+// are only set if the corresponding input bool value is set.
+//
+// INPUTS
+//     tcid  = the tcid of the transactant to load
+//      rid  - the rentable that they are tied to
+//      raf  - pointer to the flow struct to update
+//      chk  - check to see if the tcid exists in raf.People before adding.
+//             This is not always necessary, but only the caller knows.
+// isRenter  - true if we need to set the RAPerson isRenter bool to true.
+//             It should be true for Payors.
+// isOccupant- true if we need to set the RAPerson isOccupant bool to true.
+//             It should be true for Users.
+//
+// RETURNS
+//     any error encountered
+//     raf is updated
+//-----------------------------------------------------------------------------
+func addRAPtoFlow(ctx context.Context, tcid, rid int64, raf *RAFlowJSONData, chk, isRenter, isOccupant bool) error {
+	// Is this user already present?
+	if chk {
+		for l := 0; l < len(raf.People); l++ {
+			if raf.People[l].TCID == tcid {
+				if isRenter {
+					raf.People[l].IsRenter = true
+				}
+				if isOccupant {
+					raf.People[l].IsOccupant = true
+				}
+				return nil
+			}
+		}
+	}
+
+	rap, err := createRAFlowPerson(ctx, tcid, raf, isOccupant) // adds person AND associated pets and vehicles
+	if err != nil {
+		return err
+	}
+
+	if isRenter {
+		rap.IsRenter = true
+	}
+
+	if isOccupant {
+		rap.IsOccupant = true
+
+		// only tie occupants to rentable
+		var t RATiePeopleData
+		t.BID = rap.BID
+		t.TMPTCID = rap.TMPTCID
+		if rid > 0 {
+			t.PRID = rid
+		}
+		raf.Tie.People = append(raf.Tie.People, t)
+	}
+
+	// finally append in people list
+	raf.People = append(raf.People, rap)
+	return nil
+}
+
+// createRAFlowPerson returns a new RAPeopleFlowData based on the supplied
+// tcid. It does not set the Renter or Occupant flags
+//
+// INPUTS
+//          ctx  = db transaction context
+//         tcid  = the tcid of the transactant to load
+//          raf  = pointer to RAFlowJSONData
+// addDependents = adds dependents (currently pets and vehicles) to the flow
+//                 data in addition to the transactant data. The recommended
+//                 usage of this flag is to set it to true when the person
+//                 being added is a user.
+//
+// RETURNS
+//     RAPeopleFlowData structure
+//     any error encountered
+//-----------------------------------------------------------------------------
+func createRAFlowPerson(ctx context.Context, tcid int64, raf *RAFlowJSONData, addDependents bool) (RAPeopleFlowData, error) {
+	var p Transactant
+	var pu User
+	var pp Payor
+	var pr Prospect
+	var rap RAPeopleFlowData
+	var err error
+
+	raf.Meta.LastTMPTCID++
+	rap.TMPTCID = raf.Meta.LastTMPTCID // set this now so it is available when creating pets and vehicles
+	if err = GetTransactant(ctx, tcid, &p); err != nil {
+		return rap, err
+	}
+	if err = GetUser(ctx, tcid, &pu); err != nil {
+		return rap, err
+	}
+	if err = GetPayor(ctx, tcid, &pp); err != nil {
+		return rap, err
+	}
+	if err = GetProspect(ctx, tcid, &pr); err != nil {
+		return rap, err
+	}
+	MigrateStructVals(&p, &rap)
+	MigrateStructVals(&pp, &rap)
+	MigrateStructVals(&pu, &rap)
+	MigrateStructVals(&pr, &rap)
+
+	if addDependents {
+		if err = addFlowPersonVehicles(ctx, tcid, rap.TMPTCID, raf); err != nil {
+			return rap, err
+		}
+		if err = addFlowPersonPets(ctx, tcid, rap.TMPTCID, raf); err != nil {
+			return rap, err
+		}
+	}
+	return rap, nil
+}
+
+// addFlowPersonPets adds pets belonging to tcid to the supplied
+// RAFlowJSONData struct
+//
+// INPUTS
+//      ctx  = db transaction context
+//     tcid  = the tcid of the transactant to load
+//
+// RETURNS
+//     RAPetsFlowData structure
+//     any error encountered
+//-----------------------------------------------------------------------------
+func addFlowPersonPets(ctx context.Context, tcid, tmptcid int64, raf *RAFlowJSONData) error {
+	petList, err := GetPetsByTransactant(ctx, tcid)
+	if err != nil {
+		return err
+	}
+	for i := 0; i < len(petList); i++ {
+		raf.Meta.LastTMPPETID++
+		var p = RAPetsFlowData{
+			TMPTCID:  tmptcid,
+			TMPPETID: raf.Meta.LastTMPPETID,
+		}
+		MigrateStructVals(&petList[i], &p)
+		raf.Pets = append(raf.Pets, p)
+	}
+	return nil
+}
+
+// addFlowPersonVehicles adds vehicles belonging to tcid to the supplied
+// RAFlowJSONData struct
+//
+// INPUTS
+//      ctx  = db transaction context
+//     tcid  = the tcid of the transactant to load
+//
+// RETURNS
+//     RAPetsFlowData structure
+//     any error encountered
+//-----------------------------------------------------------------------------
+func addFlowPersonVehicles(ctx context.Context, tcid, tmptcid int64, raf *RAFlowJSONData) error {
+	vehicleList, err := GetVehiclesByTransactant(ctx, tcid)
+	if err != nil {
+		return err
+	}
+	for i := 0; i < len(vehicleList); i++ {
+		raf.Meta.LastTMPVID++
+		var v = RAVehiclesFlowData{
+			TMPTCID: tmptcid,
+			TMPVID:  raf.Meta.LastTMPVID,
+		}
+		MigrateStructVals(&vehicleList[i], &v)
+		raf.Vehicles = append(raf.Vehicles, v)
+	}
+	return nil
 }
