+++ ws/ra2flow.go
@@ -8,7 +8,6 @@ import (
 	"rentroll/rlib"
 	"strconv"
 	"strings"
-	"time"
 )
 
 //-------------------------------------------------------------------
@@ -348,7 +347,7 @@ func GetRA2FlowCore(ctx context.Context, ra *rlib.RentalAgreement, uid int64) (i
 	var flowID int64
 
 	// convert permanent ra to flow data and get it
-	raf, err := ConvertRA2Flow(ctx, ra)
+	raf, err := rlib.ConvertRA2Flow(ctx, ra)
 	if err != nil {
 		return flowID, err
 	}
@@ -381,388 +380,6 @@ func GetRA2FlowCore(ctx context.Context, ra *rlib.RentalAgreement, uid int64) (i
 		return flowID, err
 	}
 	return flowID, nil
-}
-
-// ConvertRA2Flow does all the heavy lifting to convert existing
-// rental agreement data to raflow data
-//
-// INPUTS:
-//     ctx    database context for transactions
-//     ra     the rental agreement to move into a flow
-//
-// RETURNS:
-//     the RAFlowJSONData
-//     any error encountered
-//-------------------------------------------------------------------------
-func ConvertRA2Flow(ctx context.Context, ra *rlib.RentalAgreement) (RAFlowJSONData, error) {
-	const funcname = "ConvertRA2Flow"
-	fmt.Printf("Entered in %s\n", funcname)
-
-	//-------------------------------------------------------------
-	// This is the datastructure we need to fill out and save...
-	//-------------------------------------------------------------
-	var raf = RAFlowJSONData{
-		Dates: RADatesFlowData{
-			BID:             ra.BID,
-			RentStart:       rlib.JSONDate(ra.RentStart),
-			RentStop:        rlib.JSONDate(ra.RentStop),
-			AgreementStart:  rlib.JSONDate(ra.AgreementStart),
-			AgreementStop:   rlib.JSONDate(ra.AgreementStop),
-			PossessionStart: rlib.JSONDate(ra.PossessionStart),
-			PossessionStop:  rlib.JSONDate(ra.PossessionStop),
-		},
-		People:      []RAPeopleFlowData{},
-		Pets:        []RAPetsFlowData{},
-		Vehicles:    []RAVehiclesFlowData{},
-		Rentables:   []RARentablesFlowData{},
-		ParentChild: []RAParentChildFlowData{},
-		Tie: RATieFlowData{
-			People: []RATiePeopleData{},
-		},
-		Meta: RAFlowMetaInfo{RAID: ra.RAID},
-	}
-
-	//-------------------------------------------------------------------------
-	// Add Users...
-	//
-	// Note: we need to add users before payors in order to
-	// ensure that all the pets and vehicles are loaded.  This is because
-	// of two behaviors of the code.  First, pets and vehicles are loaded only
-	// when the person is loaded with the User (occupant) flag set.  Second,
-	// a person is not added twice, and if you load a Payor first -- the pets
-	// and vehicles will NOT be loaded then when you call it the second time
-	// to load the same transactant as a User the code will see that the
-	// transactant has already been loaded and return without doing anything
-	// other than setting the Payor (renter) flag.
-	//-------------------------------------------------------------------------
-	n, err := rlib.GetAllRentalAgreementRentables(ctx, ra.RAID)
-	if err != nil {
-		return raf, nil
-	}
-	for j := 0; j < len(n); j++ {
-		rulist, err := rlib.GetRentableUsersInRange(ctx, n[j].RID, &ra.AgreementStart, &ra.AgreementStop)
-		if err != nil {
-			return raf, nil
-		}
-		for k := 0; k < len(rulist); k++ {
-			addRAPtoFlow(ctx, rulist[k].TCID, n[j].RID, &raf, true, false, true)
-		}
-	}
-
-	//-------------------------------------------------------------------------
-	// Add Payors...
-	//-------------------------------------------------------------------------
-	m, err := rlib.GetRentalAgreementPayorsInRange(ctx, ra.RAID, &ra.AgreementStart, &ra.AgreementStop)
-	if err != nil {
-		return raf, nil
-	}
-	for i := 0; i < len(m); i++ {
-		if err = addRAPtoFlow(ctx, m[i].TCID, 0 /*no RID here*/, &raf, true /*check dups*/, true /*renter*/, false); err != nil {
-			return raf, nil
-		}
-	}
-
-	//-------------------------------------------------------------------------
-	// Add Rentables
-	//-------------------------------------------------------------------------
-	now := time.Now()
-	o, err := rlib.GetRentalAgreementRentables(ctx, ra.RAID, &ra.AgreementStart, &ra.AgreementStop)
-	if err != nil {
-		return raf, nil
-	}
-	for i := 0; i < len(o); i++ {
-		rnt, err := rlib.GetRentable(ctx, o[i].RID)
-		if err != nil {
-			return raf, nil
-		}
-		rtr, err := rlib.GetRentableTypeRefForDate(ctx, o[i].RID, &now)
-		if err != nil {
-			return raf, nil
-		}
-		var rt rlib.RentableType
-		if err = rlib.GetRentableType(ctx, rtr.RTID, &rt); err != nil {
-			return raf, nil
-		}
-		var rfd = RARentablesFlowData{
-			BID:          o[i].BID,
-			RID:          o[i].RID,
-			RTID:         rtr.RTID,
-			RTFLAGS:      rt.FLAGS,
-			RentableName: rnt.RentableName,
-			RentCycle:    rt.RentCycle,
-		}
-
-		//---------------------------------------------------------
-		// Add the assessments associated with the Rentable...
-		// For this we want to load all 1-time fees and all
-		// recurring fees.
-		//---------------------------------------------------------
-		asms, err := rlib.GetAssessmentsByRAIDRID(ctx, rfd.BID, ra.RAID, rfd.RID)
-		if err != nil {
-			return raf, nil
-		}
-		for j := 0; j < len(asms); j++ {
-
-			//----------------------------------------------------------
-			// Get the account rule for this assessment...
-			//----------------------------------------------------------
-			ar, err := rlib.GetAR(ctx, asms[j].ARID)
-			if err != nil {
-				return raf, nil
-			}
-
-			//----------------------------------------------------------
-			// Handle Rentable Fees that are NOT Pet or Vehicle related
-			//----------------------------------------------------------
-			if ar.FLAGS&(128|256) == 0 {
-				var fee = RAFeesData{
-					ASMID:          asms[j].ASMID,
-					ARID:           asms[j].ARID,
-					ARName:         ar.Name,
-					ContractAmount: asms[j].Amount,
-					RentCycle:      asms[j].RentCycle,
-					Start:          rlib.JSONDate(asms[j].Start),
-					Stop:           rlib.JSONDate(asms[j].Stop),
-					Comment:        asms[j].Comment,
-				}
-				rfd.Fees = append(rfd.Fees, fee)
-			}
-
-			//----------------------------------------------------------
-			// Handle PET Fees
-			//----------------------------------------------------------
-			// TMPASMID        int64 // unique ID to manage fees uniquely across all fees in raflow json data
-			// ASMID           int64 // the permanent table assessment id if it is an existing RAID
-			// ARID            int64
-			// ARName          string
-			// ContractAmount  float64
-			// RentCycle       int64
-			// Start           rlib.JSONDate
-			// Stop            rlib.JSONDate
-			// AtSigningPreTax float64
-			// SalesTax        float64
-			if ar.FLAGS&(128) != 0 { // Is it a pet fee?
-				petid := asms[j].AssocElemID // find the pet...
-				for k := 0; k < len(raf.Pets); k++ {
-					if raf.Pets[k].PETID == petid {
-						raf.Meta.LastTMPASMID++
-						var pf = RAFeesData{
-							TMPASMID:       raf.Meta.LastTMPASMID,
-							ARID:           asms[j].ARID,
-							ASMID:          asms[j].ASMID,
-							ARName:         ar.Name,
-							RentCycle:      asms[j].RentCycle,
-							Start:          rlib.JSONDate(asms[j].Start),
-							Stop:           rlib.JSONDate(asms[j].Stop),
-							ContractAmount: asms[j].Amount,
-							Comment:        asms[j].Comment,
-						}
-						raf.Pets[k].Fees = append(raf.Pets[k].Fees, pf)
-						break
-					}
-				}
-			}
-			//----------------------------------------------------------
-			// Handle VEHICLE Fees
-			//----------------------------------------------------------
-			if ar.FLAGS&(256) != 0 { // Is it a vehicle fee?
-				vid := asms[j].AssocElemID // find the vehicle...
-				for k := 0; k < len(raf.Vehicles); k++ {
-					if raf.Vehicles[k].VID == vid {
-						raf.Meta.LastTMPASMID++
-						var pf = RAFeesData{
-							TMPASMID:       raf.Meta.LastTMPASMID,
-							ARID:           asms[j].ARID,
-							ASMID:          asms[j].ASMID,
-							ARName:         ar.Name,
-							ContractAmount: asms[j].Amount,
-							RentCycle:      asms[j].RentCycle,
-							Start:          rlib.JSONDate(asms[j].Start),
-							Stop:           rlib.JSONDate(asms[j].Stop),
-							Comment:        asms[j].Comment,
-						}
-						raf.Vehicles[k].Fees = append(raf.Vehicles[k].Fees, pf)
-						break
-					}
-				}
-			}
-		}
-
-		raf.Rentables = append(raf.Rentables, rfd)
-	}
-
-	return raf, nil
-}
-
-// addRAPtoFlow adds a new person to raf.People.  The renter/occupant flags
-// are only set if the corresponding input bool value is set.
-//
-// INPUTS
-//     tcid  = the tcid of the transactant to load
-//      rid  - the rentable that they are tied to
-//      raf  - pointer to the flow struct to update
-//      chk  - check to see if the tcid exists in raf.People before adding.
-//             This is not always necessary, but only the caller knows.
-// isRenter  - true if we need to set the RAPerson isRenter bool to true.
-//             It should be true for Payors.
-// isOccupant- true if we need to set the RAPerson isOccupant bool to true.
-//             It should be true for Users.
-//
-// RETURNS
-//     any error encountered
-//     raf is updated
-//-----------------------------------------------------------------------------
-func addRAPtoFlow(ctx context.Context, tcid, rid int64, raf *RAFlowJSONData, chk, isRenter, isOccupant bool) error {
-	// Is this user already present?
-	if chk {
-		for l := 0; l < len(raf.People); l++ {
-			if raf.People[l].TCID == tcid {
-				if isRenter {
-					raf.People[l].IsRenter = true
-				}
-				if isOccupant {
-					raf.People[l].IsOccupant = true
-				}
-				return nil
-			}
-		}
-	}
-
-	rap, err := createRAFlowPerson(ctx, tcid, raf, isOccupant) // adds person AND associated pets and vehicles
-	if err != nil {
-		return err
-	}
-
-	if isRenter {
-		rap.IsRenter = true
-	}
-
-	if isOccupant {
-		rap.IsOccupant = true
-
-		// only tie occupants to rentable
-		var t RATiePeopleData
-		t.BID = rap.BID
-		t.TMPTCID = rap.TMPTCID
-		if rid > 0 {
-			t.PRID = rid
-		}
-		raf.Tie.People = append(raf.Tie.People, t)
-	}
-
-	// finally append in people list
-	raf.People = append(raf.People, rap)
-	return nil
-}
-
-// createRAFlowPerson returns a new RAPeopleFlowData based on the supplied
-// tcid. It does not set the Renter or Occupant flags
-//
-// INPUTS
-//          ctx  = db transaction context
-//         tcid  = the tcid of the transactant to load
-//          raf  = pointer to RAFlowJSONData
-// addDependents = adds dependents (currently pets and vehicles) to the flow
-//                 data in addition to the transactant data. The recommended
-//                 usage of this flag is to set it to true when the person
-//                 being added is a user.
-//
-// RETURNS
-//     RAPeopleFlowData structure
-//     any error encountered
-//-----------------------------------------------------------------------------
-func createRAFlowPerson(ctx context.Context, tcid int64, raf *RAFlowJSONData, addDependents bool) (RAPeopleFlowData, error) {
-	var p rlib.Transactant
-	var pu rlib.User
-	var pp rlib.Payor
-	var pr rlib.Prospect
-	var rap RAPeopleFlowData
-	var err error
-
-	raf.Meta.LastTMPTCID++
-	rap.TMPTCID = raf.Meta.LastTMPTCID // set this now so it is available when creating pets and vehicles
-	if err = rlib.GetTransactant(ctx, tcid, &p); err != nil {
-		return rap, err
-	}
-	if err = rlib.GetUser(ctx, tcid, &pu); err != nil {
-		return rap, err
-	}
-	if err = rlib.GetPayor(ctx, tcid, &pp); err != nil {
-		return rap, err
-	}
-	if err = rlib.GetProspect(ctx, tcid, &pr); err != nil {
-		return rap, err
-	}
-	rlib.MigrateStructVals(&p, &rap)
-	rlib.MigrateStructVals(&pp, &rap)
-	rlib.MigrateStructVals(&pu, &rap)
-	rlib.MigrateStructVals(&pr, &rap)
-
-	if addDependents {
-		if err = addFlowPersonVehicles(ctx, tcid, rap.TMPTCID, raf); err != nil {
-			return rap, err
-		}
-		if err = addFlowPersonPets(ctx, tcid, rap.TMPTCID, raf); err != nil {
-			return rap, err
-		}
-	}
-	return rap, nil
-}
-
-// addFlowPersonPets adds pets belonging to tcid to the supplied
-// RAFlowJSONData struct
-//
-// INPUTS
-//      ctx  = db transaction context
-//     tcid  = the tcid of the transactant to load
-//
-// RETURNS
-//     RAPetsFlowData structure
-//     any error encountered
-//-----------------------------------------------------------------------------
-func addFlowPersonPets(ctx context.Context, tcid, tmptcid int64, raf *RAFlowJSONData) error {
-	petList, err := rlib.GetPetsByTransactant(ctx, tcid)
-	if err != nil {
-		return err
-	}
-	for i := 0; i < len(petList); i++ {
-		raf.Meta.LastTMPPETID++
-		var p = RAPetsFlowData{
-			TMPTCID:  tmptcid,
-			TMPPETID: raf.Meta.LastTMPPETID,
-		}
-		rlib.MigrateStructVals(&petList[i], &p)
-		raf.Pets = append(raf.Pets, p)
-	}
-	return nil
-}
-
-// addFlowPersonVehicles adds vehicles belonging to tcid to the supplied
-// RAFlowJSONData struct
-//
-// INPUTS
-//      ctx  = db transaction context
-//     tcid  = the tcid of the transactant to load
-//
-// RETURNS
-//     RAPetsFlowData structure
-//     any error encountered
-//-----------------------------------------------------------------------------
-func addFlowPersonVehicles(ctx context.Context, tcid, tmptcid int64, raf *RAFlowJSONData) error {
-	vehicleList, err := rlib.GetVehiclesByTransactant(ctx, tcid)
-	if err != nil {
-		return err
-	}
-	for i := 0; i < len(vehicleList); i++ {
-		raf.Meta.LastTMPVID++
-		var v = RAVehiclesFlowData{
-			TMPTCID: tmptcid,
-			TMPVID:  raf.Meta.LastTMPVID,
-		}
-		rlib.MigrateStructVals(&vehicleList[i], &v)
-		raf.Vehicles = append(raf.Vehicles, v)
-	}
-	return nil
 }
 
 // wsdoc {
